backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet>=3.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Async Task Queue
celery==5.3.6
redis==5.0.1

# File Processing
PyMuPDF==1.23.8
pandas==2.1.4

# AI Integration (Google Gemini)
google-generativeai==0.3.2

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Linting & Type Checking
ruff==0.1.14
mypy==1.8.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
